refactor(MenuProduct): menuProducts 일급컬렉션 적용 및 amount 계산 로직 이동
