+++ src/test/kotlin/kitchenpos/menu/tobe/domain/MenuTest.kt
@@ -1,29 +1,18 @@
 package kitchenpos.menu.tobe.domain
 
-import java.math.BigDecimal
 import kitchenpos.utils.Fixtures
-import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
 
 class MenuTest {
+
     @Test
-    @DisplayName("`Menu`의 `amount`를 계산한다")
-    fun amount() {
+    @DisplayName("Menu를 생성한다")
+    fun create() {
+        // given
+        val 추천메뉴 = Fixtures.menuGroup(name = "추천메뉴")
+
         val 양념치킨 = Fixtures.product(name = "양념치킨", price = 16000)
         val 후라이드치킨 = Fixtures.product(name = "후라이드치킨", price = 17000)
-        val menu = Fixtures.menu(
-            name = "후라이드양념치킨",
-            price = 0,
-            displayed = true,
-            menuProducts = listOf(
-                Fixtures.menuProduct(product = 양념치킨, quantity = 1),
-                Fixtures.menuProduct(product = 후라이드치킨, quantity = 2)
-            )
-        )
-
-        val amount = menu.amount()
-
-        assertThat(amount).isEqualTo(BigDecimal.valueOf(50000))
     }
 }
