+++ src/test/kotlin/kitchenpos/menu/tobe/infra/FakeMenuRepository.kt
@@ -26,6 +26,6 @@ class FakeMenuRepository(
     }
 
     override fun findAllByProductId(productId: UUID): List<Menu> {
-        return menus.values.filter { it.menuProducts.any { it.product.id == productId } }
+        return menus.values.filter { it.menuProducts.menuProducts.any { it.productId == productId } }
     }
 }
